Move EtasMemberOverlapReport to lib/report
