+++ spec/report/etas_member_overlap_report_spec.rb
@@ -2,9 +2,9 @@
 
 require "spec_helper"
 require "pp"
-require_relative "../../bin/reports/export_etas_overlap_report"
+require "report/etas_member_overlap_report"
 
-RSpec.describe EtasMemberOverlapReport do
+RSpec.describe Report::EtasMemberOverlapReport do
   let(:h) { build(:holding) }
   let(:h2) { build(:holding, organization: "ualberta") }
   let(:ht) { build(:ht_item, ocns: [h.ocn], access: "deny") }
